backend/requirements.txt
# FinAI Backend - Production Requirements
# Install: pip install -r requirements.txt

# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
gunicorn==23.0.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
aiosqlite==0.20.0            # async SQLite
asyncpg==0.30.0              # async PostgreSQL
psycopg2-binary==2.9.10      # PostgreSQL (production)

# Data Processing
pandas==2.2.3
openpyxl==3.1.5
xlsxwriter==3.2.0
numpy==2.1.3

# AI
anthropic==0.39.0
llama-index==0.12.3
# llama-index-vector-stores-chroma commented out: conflicts with llama-index-core
# llama-index-vector-stores-chroma==0.3.0

# Vector Store (standalone — used without llama-index integration layer)
chromadb>=0.5.0              # 1.x is fine; the <1.0.0 constraint was overly strict

# File Handling
python-multipart==0.0.12
aiofiles==24.1.0

# Document Intelligence (PDF / Word / Excel ingestion for RAG pipeline)
pypdf>=4.0.0                 # PDF text extraction (already installed as pypdf 5.x)
python-docx>=1.1.0           # Word document (.docx) parsing (already installed)
# openpyxl already listed above for Excel

# Config & Security
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4       # kept for backward compat; auth.py uses bcrypt directly
bcrypt>=4.0.0                # direct bcrypt usage to avoid passlib 1.7.4 incompatibility
python-jose[cryptography]==3.3.0
email-validator>=2.0.0       # Required for pydantic EmailStr

# HTTP Client
httpx==0.28.1

# Utilities
python-dateutil==2.9.0
pytz==2024.2
orjson==3.10.12

# Logging — structured JSON (see app/logging_config.py)
python-json-logger>=2.0.7       # JSON log formatting, graceful fallback in logging_config.py

# Caching (optional — app works without it, falls back to in-memory)
redis[hiredis]==5.2.1

pytest==8.3.4
pytest-asyncio==0.24.0
httpx>=0.28.0                    # For FastAPI TestClient in integration tests
anyio[trio]>=4.0.0               # Required by pytest-asyncio

# Telemetry / Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.40b0
opentelemetry-instrumentation-sqlalchemy>=0.40b0
opentelemetry-exporter-otlp>=1.20.0
